revisão de funções

+++ include/directory.hpp
@@ -11,10 +11,10 @@
 bool create_directory(const std::string& dirname, uint16_t parent_cluster);
 
 // Lista todas as entradas de um diretório
-std::vector<DirEntry> list_directory(int cluster_index);
+std::vector<DirEntry> list_directory(uint16_t cluster_index);
 
 // Procura uma entrada de diretório por nome
-std::optional<DirEntry> find_entry(const std::string& name, int cluster_index);
+std::optional<DirEntry> find_entry(const std::string& name, uint16_t cluster_index);
 
 // Adiciona uma nova entrada em um diretório
 bool add_entry_to_directory(const DirEntry& entry, uint16_t cluster_index);
